refactor: apply pypots.vec.ts2vec.core._TS2Vec directly in classification TS2Vec;
